+++ src/services/mod.rs
@@ -34,6 +34,7 @@ use crate::cache::Caches;
 use crate::config::Config;
 use crate::utils::futures::ThreadPool;
 
+pub mod bitcode;
 pub mod cacher;
 pub mod cficaches;
 pub mod download;
@@ -41,6 +42,7 @@ pub mod objects;
 pub mod symbolication;
 pub mod symcaches;
 
+use self::bitcode::BitcodeService;
 use self::cficaches::CfiCacheActor;
 use self::download::DownloadService;
 use self::objects::ObjectsActor;
@@ -74,7 +76,9 @@ impl Service {
             .clear_tmp(&config)
             .context("failed to clear tmp caches")?;
         let objects = ObjectsActor::new(caches.object_meta, caches.objects, downloader.clone());
-        let symcaches = SymCacheActor::new(caches.symcaches, objects.clone(), cpu_pool.clone());
+        let bitcode = BitcodeService::new(caches.auxdifs, downloader.clone());
+        let symcaches =
+            SymCacheActor::new(caches.symcaches, objects.clone(), bitcode, cpu_pool.clone());
         let cficaches = CfiCacheActor::new(caches.cficaches, objects.clone(), cpu_pool.clone());
 
         let symbolication = SymbolicationActor::new(