+++ src/utils/mod.rs
@@ -1,4 +1,5 @@
 pub mod addr;
+pub mod compression;
 pub mod futures;
 pub mod hex;
 pub mod http;