+++ symsorter/src/main.rs
@@ -7,6 +7,7 @@ mod utils;
 
 mod app;
 mod config;
+mod difs;
 
 fn main() {
     app::main();